update to latest language-sally, fix tests
This commit updates to the changes in language-sally 0.2.1.0.

In particular, queries now receive a state they can use.

This commit also updates the test suite (which is not actually testing
anything but running the translation on examples), to account for all
recent changes.

This commit also removes some code that has made obsolete by the same
recent changes: we no longer need to generate accessors as we receive a
What4.SymStruct that we can reach into using What4.structField.